dbd/requirements.txt
sortedcontainers>=2.4.0
prompt-toolkit>=3.0.38
pyzmq>=25.1.1
tabulate>=0.9.0
rapidyaml>=0.5.0
deprecation>=2.1.0
docopt>=0.6.2
pyelftools>=0.30
fuzzywuzzy>=0.18.0
python-levenshtein>=0.22.0
pyyaml>=5.3.1